Add SDLPort for mac
git-svn-id: https://nekonme.googlecode.com/svn/trunk@238 1509560c-5e2a-0410-865c-31c25e1cfdef

+++ project2/windows/SimpleFrame.cpp
@@ -163,7 +163,7 @@ public:
    HWND         mHWND;
    HDC          mDC;
    uint32       mFlags;
-   int          mFrameRate;
+   double       mFrameRate;
    EventHandler mHandler;
    DIBSurface   *mBMP;
    void         *mHandlerData;
